// code/frontend/src/app/services/models/cadastral-filter-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CadastralFilterRequest {
  maxEnergyClass: number;
  maxFloor: number;
  maxPrice: number;
  maxRooms: number;
  maxSquareMeters: number;
  minEnergyClass: number;
  minFloor: number;
  minPrice: number;
  minRooms: number;
  minSquareMeters: number;
}